+++ tests/php/unit/JsonLdReturnPolicyTest.php
@@ -45,6 +45,9 @@ class JsonLdReturnPolicyTest extends \PHPUnit\Framework\TestCase {
 				return $url . $sep . $query . $fragment;
 			}
 		);
+		Functions\when( 'home_url' )->alias(
+			static fn( $path = '' ) => 'https://example.com' . $path
+		);
 		Functions\when( 'wc_get_product_cat_ids' )->justReturn( [] );
 		Functions\when( 'wc_get_base_location' )->justReturn(
 			[ 'country' => 'US' ]
@@ -103,6 +106,8 @@ class JsonLdReturnPolicyTest extends \PHPUnit\Framework\TestCase {
 		$product->shouldReceive( 'has_dimensions' )->andReturn( false );
 		$product->shouldReceive( 'get_dimensions' )->andReturn( [] );
 		$product->shouldReceive( 'get_attributes' )->andReturn( [] );
+		$product->shouldReceive( 'get_children' )->andReturn( [] );
+		$product->shouldReceive( 'get_sku' )->andReturn( '' );
 		return $product;
 	}